# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/vehicles": {
      "post": {
        "operationId": "VehiclesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Veículo criado.",
            "headers": {
              "Location": {
                "description": "URL do recurso criado",
                "schema": {
                  "type": "string",
                  "example": "/api/customers/{id}"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (validação do DTO, UUID malformado ou campo não permitido).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Papel sem permissão, ou recurso de outro cliente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflito com um recurso existente (unicidade).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requisições excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Cadastra um veículo (admin, analyst)",
        "tags": [
          "vehicles"
        ]
      },
      "get": {
        "operationId": "VehiclesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Página de veículos."
          },
          "400": {
            "description": "Entrada inválida (validação do DTO, UUID malformado ou campo não permitido).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Papel sem permissão, ou recurso de outro cliente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requisições excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Lista veículos, paginado (admin, analyst)",
        "tags": [
          "vehicles"
        ]
      }
    },
    "/api/vehicles/{id}": {
      "get": {
        "description": "A representação é escolhida pelo header Accept (application/json é o padrão; application/xml devolve a ficha técnica em XML). customer só vê os próprios veículos.",
        "operationId": "VehiclesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (validação do DTO, UUID malformado ou campo não permitido).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Papel sem permissão, ou recurso de outro cliente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "406": {
            "description": "Formato pedido no header Accept não é suportado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requisições excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Detalha um veículo em JSON ou XML",
        "tags": [
          "vehicles"
        ]
      },
      "patch": {
        "operationId": "VehiclesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (validação do DTO, UUID malformado ou campo não permitido).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Papel sem permissão, ou recurso de outro cliente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflito com um recurso existente (unicidade).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requisições excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Atualiza parcialmente um veículo (admin, analyst)",
        "tags": [
          "vehicles"
        ]
      },
      "delete": {
        "operationId": "VehiclesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Veículo removido."
          },
          "400": {
            "description": "Entrada inválida (validação do DTO, UUID malformado ou campo não permitido).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Papel sem permissão, ou recurso de outro cliente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requisições excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Remove um veículo (admin)",
        "tags": [
          "vehicles"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ford Conecta — API (Web Services)",
    "description": "Camada de Serviços REST da plataforma Ford Conecta — FIAP Ford 2026, Desafio 02.\n\n**Autenticação:** faça `POST /api/auth/login`, copie o `accessToken` e clique em **Authorize**. Rotas sem cadeado são públicas.\n\n**Papéis:** `admin` (tudo), `analyst` (pós-venda da concessionária) e `customer` (dono do veículo, só os próprios dados).\n\n**Erros:** toda resposta de erro segue o schema `ErrorResponseDto`.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          },
          "message": {
            "oneOf": [
              {
                "type": "string",
                "example": "Cliente não encontrado"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "email must be an email"
                ]
              }
            ],
            "description": "Mensagem, ou lista de mensagens quando a validação de entrada falha"
          },
          "path": {
            "type": "string",
            "example": "/api/customers/00000000-0000-0000-0000-000000000000"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-26T12:00:00.000Z"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "timestamp"
        ]
      },
...
      "Customer": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "document": {
            "type": "string",
            "description": "CPF (somente dígitos). Identificador único do cliente."
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "vehicles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vehicle"
            }
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "name",
          "document",
          "email",
          "phone",
          "city",
          "state",
          "vehicles",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "Dealership": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "vehicles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vehicle"
            }
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "name",
          "city",
          "region",
          "vehicles",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "Vehicle": {
        "type": "object",
        "properties": {
          "vin": {
            "type": "string",
            "description": "Vehicle Identification Number — 17 caracteres."
          },
          "model": {
            "type": "string"
          },
          "modelYear": {
            "type": "number"
          },
          "purchaseDate": {
            "type": "string"
          },
          "odometer": {
            "type": "number",
            "description": "Hodômetro atual (km) — dado de veículo conectado (IoT). NÃO é usado na predição."
          },
          "customerId": {
            "type": "string"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "dealershipId": {
            "type": "string"
          },
          "dealership": {
            "$ref": "#/components/schemas/Dealership"
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "vin",
          "model",
          "modelYear",
          "purchaseDate",
          "odometer",
          "customerId",
          "customer",
          "dealershipId",
          "dealership",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateVehicleDto": {
        "type": "object",
        "properties": {
          "vin": {
            "type": "string",
            "maxLength": 17,
            "example": "9BFZK54P5J8000001",
            "description": "VIN (até 17 caracteres)"
          },
          "model": {
            "type": "string",
            "maxLength": 60,
            "example": "Ranger"
          },
          "modelYear": {
            "type": "number",
            "minimum": 1990,
            "maximum": 2100,
            "example": 2024
          },
          "purchaseDate": {
            "type": "string",
            "example": "2024-03-15",
            "description": "Data da compra (YYYY-MM-DD)"
          },
          "odometer": {
            "type": "number",
            "minimum": 0,
            "example": 12000,
            "default": 0,
            "description": "Hodômetro (km)"
          },
          "customerId": {
            "type": "string",
            "format": "uuid",
            "description": "ID (UUID) do cliente dono do veículo"
          },
          "dealershipId": {
            "type": "string",
            "format": "uuid",
            "description": "ID (UUID) da concessionária da compra"
          }
        },
        "required": [
          "vin",
          "model",
          "modelYear",
          "purchaseDate",
          "customerId"
        ]
      },
      "UpdateVehicleDto": {
        "type": "object",
        "properties": {
          "vin": {
            "type": "string",
            "maxLength": 17,
            "example": "9BFZK54P5J8000001",
            "description": "VIN (até 17 caracteres)"
          },
          "model": {
            "type": "string",
            "maxLength": 60,
            "example": "Ranger"
          },
          "modelYear": {
            "type": "number",
            "minimum": 1990,
            "maximum": 2100,
            "example": 2024
          },
          "purchaseDate": {
            "type": "string",
            "example": "2024-03-15",
            "description": "Data da compra (YYYY-MM-DD)"
          },
          "odometer": {
            "type": "number",
            "minimum": 0,
            "example": 12000,
            "default": 0,
            "description": "Hodômetro (km)"
          },
          "dealershipId": {
            "type": "string",
            "format": "uuid",
            "description": "ID (UUID) da concessionária da compra"
          }
        }
      },
...
    }
  }
}